requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
streamlit
python-dotenv
pypdf
